+++ src/main/java/com/smunity/api/domain/question/dto/AnswerDto.java
@@ -1,8 +1,8 @@
 package com.smunity.api.domain.question.dto;
 
 import com.smunity.api.domain.account.entity.User;
-import com.smunity.api.domain.question.domain.Answer;
-import com.smunity.api.domain.question.domain.Question;
+import com.smunity.api.domain.question.entity.Answer;
+import com.smunity.api.domain.question.entity.Question;
 import lombok.Builder;
 import lombok.Data;
 import java.time.LocalDateTime;
