+++ src/main/java/com/smunity/api/domain/question/entity/Answer.java
@@ -1,4 +1,4 @@
-package com.smunity.api.domain.question.domain;
+package com.smunity.api.domain.question.entity;
 
 import com.smunity.api.domain.account.entity.User;
 import com.smunity.api.global.common.BaseEntity;
@@ -11,23 +11,22 @@ import javax.persistence.*;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "qna_question")
-public class Question extends BaseEntity {
+@Table(name = "qna_answer")
+public class Answer extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(nullable = false)
-    private String subject;
-
-    @Column(nullable = false)
     private String content;
-
-    @Column(nullable = false)
-    private Boolean anonymous;
 
     @ManyToOne
     @JoinColumn(name = "author_id")
     @ToString.Exclude
     private User author;
+
+    @OneToOne
+    @JoinColumn(name = "question_id")
+    @ToString.Exclude
+    private Question question;
 }
