+++ src/main/java/com/smunity/api/domain/question/repository/AnswerRepository.java
@@ -1,6 +1,6 @@
 package com.smunity.api.domain.question.repository;
 
-import com.smunity.api.domain.question.domain.Answer;
+import com.smunity.api.domain.question.entity.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
