+++ src/main/java/com/smunity/api/domain/question/service/impl/AnswerServiceImpl.java
@@ -2,8 +2,8 @@ package com.smunity.api.domain.question.service.impl;
 
 import com.smunity.api.domain.account.entity.User;
 import com.smunity.api.domain.account.repository.UserRepository;
-import com.smunity.api.domain.question.domain.Answer;
-import com.smunity.api.domain.question.domain.Question;
+import com.smunity.api.domain.question.entity.Answer;
+import com.smunity.api.domain.question.entity.Question;
 import com.smunity.api.domain.question.dto.AnswerDto;
 import com.smunity.api.domain.question.repository.AnswerRepository;
 import com.smunity.api.domain.question.repository.QuestionRepository;
